+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 **Fixes:**
 - Fixed custom YAML installer issue with different image (Issue [#613](https://github.com/NetApp/trident/issues/613)).
 - Fixed snapshot size calculation (Issue [#611](https://github.com/NetApp/trident/issues/611)).
+- Fixed issue where clones of XFS volumes could not be mounted on the same node as the source volume (Issue [#514](https://github.com/NetApp/trident/issues/514)).
 
 **Enhancements:**
 
@@ -14,7 +15,7 @@
 
 **Deprecations:**
 
-## v21.04.0
+## v21.07.0
 
 - **IMPORTANT**: Trident has updated its recommendations for the iSCSI setup on worker nodes. Please carefully
   read the ``Preparing the worker node`` section of the documentation. Please ensure worker node multipathing